Auth: move to test package
