+++ pkg/services/auth/authtest/testing.go
@@ -1,4 +1,4 @@
-package auth
+package authtest
 
 import (
 	"context"
